Setup tooling and dependencies
This PR adds TypeScript, linting, formatting and emotion for styling.
