+++ .gitignore
@@ -22,4 +22,5 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-.idea
+.idea.vscode
+.vscode/
