+++ .prettierignore
@@ -1,0 +1,7 @@
+package.json
+package-lock.json
+webpack.*
+gulpfile.*
+eslintrc.*
+*.json
+puppeteer.*
