+++ package.json
@@ -1,39 +1,59 @@
 {
-  "name": "ownemployed",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "antd": "^4.1.4",
-    "history": "^4.10.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router": "^5.1.2",
-    "react-router-config": "^5.1.1",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.1"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
+    "name": "ownemployed",
+    "version": "0.1.0",
+    "private": true,
+    "dependencies": {
+        "@testing-library/jest-dom": "^4.2.4",
+        "@testing-library/react": "^9.3.2",
+        "@testing-library/user-event": "^7.1.2",
+        "@types/react-router-config": "^5.0.1",
+        "@types/react-router-dom": "^5.1.5",
+        "antd": "^4.1.4",
+        "emotion": "^10.0.27",
+        "history": "^4.10.1",
+        "react": "^16.13.1",
+        "react-dom": "^16.13.1",
+        "react-router": "^5.1.2",
+        "react-router-config": "^5.1.1",
+        "react-router-dom": "^5.1.2",
+        "react-scripts": "3.4.1",
+        "typescript": "~3.7.2"
+    },
+    "devDependencies": {
+        "eslint": "^6.8.0",
+        "husky": "^4.2.5",
+        "lint-staged": "^10.2.2",
+        "mrm": "^2.3.0",
+        "prettier": "^2.0.5"
+    },
+    "scripts": {
+        "start": "BROWSER=none react-scripts start",
+        "build": "react-scripts build",
+        "test": "react-scripts test",
+        "eject": "react-scripts eject"
+    },
+    "husky": {
+        "hooks": {
+            "pre-commit": "lint-staged"
+        }
+    },
+    "lint-staged": {
+        "*.{js,ts,jsx,tsx}": "eslint --fix",
+        "*.{js,jsx,ts,tsx,md,html,css}": "prettier --write"
+    },
+    "eslintConfig": {
+        "extends": "react-app"
+    },
+    "browserslist": {
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
+    }
 }
