+++ src/App.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React from 'react'
 import { Router } from 'react-router-dom'
 import { renderRoutes } from 'react-router-config'
 import { createBrowserHistory } from 'history'
@@ -9,11 +9,7 @@ import routes from './routes'
 const history = createBrowserHistory()
 
 function App() {
-  return (
-    <Router history={history}>
-      {renderRoutes(routes)}
-    </Router>
-  );
+    return <Router history={history}>{renderRoutes(routes)}</Router>
 }
 
-export default App;
+export default App
